Gather facts again in Homeshick role
This role needs to be run as an unprivileged user so that the repos that it creates are owned by the unprivileged user, not root.

This was happening, but the `ansible_user_dir` is not refreshed when `become: false` is passed with the role.

The `setup` module needs to be run again after dropping back to the unprivileged user to refresh the fact.